리팩토링 완료
