+++ src/pages/detail/detail.scss
@@ -1,5 +1,4 @@
 @import url(//spoqa.github.io/spoqa-han-sans/css/SpoqaHanSans-kr.css);
-@import '../../styles/mixin';
 
 * {
 	box-sizing: border-box;
