+++ src/pages/detail/detailLikebutton/detailLikeButton.js
